Refactor Render component
