+++ mathesar_ui/src/component-library/index.ts
@@ -35,6 +35,7 @@ export { default as PasswordInput } from './password-input/PasswordInput.svelte'
 export { default as Progress } from './progress/Progress.svelte';
 export { default as Radio } from './radio/Radio.svelte';
 export { default as RadioGroup } from './radio-group/RadioGroup.svelte';
+export { default as Render } from './render/Render.svelte';
 export { default as Skeleton } from './skeleton/Skeleton.svelte';
 export { default as Spinner } from './spinner/Spinner.svelte';
 export { default as SpinnerArea } from './spinner-area/SpinnerArea.svelte';
