+++ mathesar_ui/src/component-library/render/Render.svelte
@@ -1,17 +1,32 @@
 <script lang="ts">
   import type { SvelteComponent } from 'svelte';
 
-  import type { ComponentWithProps } from '../types';
+  import MarginTrim from '@mathesar-component-library-dir/margin-trim/MarginTrim.svelte';
+  import type { ComponentWithProps } from '@mathesar-component-library-dir/types';
+  import RenderComponentWithProps from './RenderComponentWithProps.svelte';
 
   type T = $$Generic<SvelteComponent>;
 
-  export let componentWithProps: ComponentWithProps<T>;
-
-  // I'm not sure how to fix the use of `any` here. Suggestions welcome!
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  $: component = componentWithProps.component as any;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  $: props = componentWithProps.props as any;
+  export let arg: string | string[] | ComponentWithProps<T> | undefined =
+    undefined;
 </script>
 
-<svelte:component this={component} {...props} />
+{#if arg === undefined}
+  {''}
+{:else if typeof arg === 'string'}
+  {arg}
+{:else if Array.isArray(arg)}
+  {#if arg.length === 0}
+    {''}
+  {:else if arg.length === 1}
+    {arg[0]}
+  {:else}
+    <MarginTrim>
+      {#each arg as paragraph}
+        <p>{paragraph}</p>
+      {/each}
+    </MarginTrim>
+  {/if}
+{:else}
+  <RenderComponentWithProps componentWithProps={arg} />
+{/if}
