+++ mathesar_ui/src/component-library/string-or-component/StringOrComponent.svelte
@@ -1,9 +1,8 @@
 <!--
   @component
 
-  @deprecated in favor of `StringOrComponentTyped` which uses
-  `ComponentWithProps` for better type safety in comparing a component to its
-  props.
+  @deprecated in favor of `Render` which accepts `ComponentWithProps` for better
+  type safety in comparing a component to its props.
 -->
 <script lang="ts">
   import type { ComponentAndProps } from '@mathesar-component-library-dir/types';
