--- mathesar_ui/src/component-library/string-or-component/StringOrComponentTyped.svelte
@@ -1,25 +1,0 @@
-<!--
-  @component
-
-  This component replaces `StringOrComponent` (which is deprecated)
--->
-<script lang="ts">
-  import type { SvelteComponent } from 'svelte';
-
-  import type { ComponentWithProps } from '@mathesar-component-library-dir/types';
-  import Render from '../render/Render.svelte';
-
-  type T = $$Generic<SvelteComponent>;
-
-  export let arg: string | string[] | ComponentWithProps<T>;
-</script>
-
-{#if typeof arg === 'string'}
-  {arg}
-{:else if Array.isArray(arg)}
-  {#each arg as paragraph}
-    <p>{paragraph}</p>
-  {/each}
-{:else}
-  <Render componentWithProps={arg} />
-{/if}
